refactor: split CourseModal into smaller components (#1739)
