+++ frontend/src/components/CourseModal/CourseModal.module.css
@@ -1,31 +1,3 @@
-.backArrow {
-  height: 30px;
-}
-
-.badges {
-  height: auto;
-  display: flex;
-  align-items: center;
-  margin-top: 0.25rem;
-}
-
-.courseCodes {
-  height: auto;
-  font-size: 16px;
-  line-height: 20px;
-  margin: 0 0.5rem 0 0;
-}
-
-.crossListingLink,
-.crossListingLink:hover {
-  color: inherit;
-  text-decoration: underline;
-}
-
-.cancelledText {
-  color: rgb(255 98 98);
-}
-
 .dialog {
   max-width: 800px !important;
 }
@@ -64,102 +36,7 @@
   color: var(--color-text-secondary);
 }
 
-.modalTop {
-  width: calc(100% - 1rem);
-  display: flex;
-  align-items: center;
-  gap: 1em;
-  margin-top: 0.5rem;
-  margin-left: 0.5rem;
-}
-
-.modalTitle {
-  font-weight: 550;
-  line-height: 1.2 !important;
-}
-
-.modalControls {
-  width: calc(100% - 1rem);
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap-reverse;
-  margin-left: 0.5rem;
-}
-
-.tabs {
-  width: 220px;
-  display: flex;
-  align-items: end;
-}
-
-.tabButton {
-  width: 50%;
-  padding: 0.5em 0;
-  border: none;
-  border-radius: 8px 8px 0 0;
-  background-color: var(--color-surface);
-  color: var(--color-text);
-  font-size: 16px;
-  font-weight: 500;
-}
-
-.tabButton.tabSelected {
-  background-color: var(--color-primary-subdued);
-  color: var(--color-text);
-}
-
-.shareButton {
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-}
-
-.moreDropdown {
-  cursor: pointer;
-  margin-top: auto;
-  margin-bottom: auto;
-  padding: 0;
-}
-
-.moreDropdown svg,
-.shareButton svg {
-  color: var(--color-primary);
-}
-
-.moreDropdown:hover svg,
-.shareButton:hover svg {
-  color: var(--color-primary-hover);
-}
-
-.moreDropdown :global(.dropdown-toggle) {
-  padding: 0;
-}
-
-.moreDropdown :global(.dropdown-toggle)::after {
-  display: none;
-}
-
-.moreDropdown :global(.dropdown-menu) {
-  width: 250px;
-}
-
-.toolBar {
-  display: flex;
-  gap: 1em;
-}
-
 @media (max-width: calc(768px - 0.5px)) {
-  .modalTop {
-    flex-direction: column;
-    align-items: start;
-    gap: 0;
-    margin: 0.5rem;
-  }
-
-  .modalTitle {
-    line-height: 1 !important;
-    font-size: 18px !important;
-  }
-
   .dialog :global(.modal-dialog) {
     margin: 1.75rem 0.5rem;
   }
