+++ frontend/src/components/CourseModal/EvaluationsPanel/EvaluationRatings.module.css
