+++ frontend/src/components/CourseModal/EvaluationsPanel/EvaluationRatings.tsx
@@ -1,7 +1,7 @@
 import RatingsGraph from './RatingsGraph';
-import type { SearchEvaluationNarrativesQuery } from '../../generated/graphql-types';
-import { evalQuestionTags } from '../../utilities/constants';
-import { TextComponent } from '../Typography';
+import type { SearchEvaluationNarrativesQuery } from '../../../generated/graphql-types';
+import { evalQuestionTags } from '../../../utilities/constants';
+import { TextComponent } from '../../Typography';
 import styles from './EvaluationRatings.module.css';
 
 const tagIndex = Object.fromEntries(
