+++ frontend/src/components/CourseModal/EvaluationsPanel/RatingsGraph.module.css
