+++ frontend/src/components/CourseModal/OverviewPanel/OverviewInfo.module.css
