+++ frontend/src/components/CourseModal/OverviewPanel/OverviewRatings.module.css
